recfactor: AuthController remove
- 필요없는 컨트롤러 제거
